Extracted out the extractUserInfo and extractExhibitionInfo to the utils file.
